+++ meaningless/utilities/common.py
@@ -5,7 +5,7 @@ import re
 
 # This is a collection of helper methods used across the various modules.
 
-MEANINGLESS_VERSION = '1.3.0'
+MEANINGLESS_VERSION = '1.4.0'
 '''
 The current version of the Meaningless library.
 '''
@@ -660,10 +660,6 @@ BIBLE_TRANSLATIONS = {
     'NLV': {
         **get_bible_data_for_language('English'),
         BIBLE_KEY_COPYRIGHT: 'https://www.biblegateway.com/versions/New-Life-Version-NLV-Bible/#copy'
-    },
-    'NMB': {
-        **get_bible_data_for_language('English', mode=1),
-        BIBLE_KEY_COPYRIGHT: 'https://www.biblegateway.com/versions/New-Matthew-Bible-NMB/#copy'
     },
     'NOG': {
         **get_bible_data_for_language('English'),