+++ test/unit_tests_bible_base_downloader.py
@@ -263,7 +263,7 @@ class UnitTests(unittest.TestCase):
         self.assertRaises(UnsupportedTranslationError, bible.download_book, 'Ecclesiastes')
 
     def test_base_download_on_translation_with_versenum_tags(self):
-        download_path = './tmp/test_base_download_with_valid_empty_passage/'
+        download_path = './tmp/test_base_download_on_translation_with_versenum_tags/'
         bible = BaseDownloader(file_writing_function=yaml_file_interface.write, default_directory=download_path,
                                translation='NIVUK')
         bible.download_passage('Mark', 9, 44)
@@ -271,6 +271,28 @@ class UnitTests(unittest.TestCase):
         expected_contents = '⁴⁴'
         self.assertEqual(len(downloaded_file['Mark'][9]), 1, 'Incorrect number of passages downloaded')
         self.assertEqual(downloaded_file['Mark'][9][44], expected_contents, 'Passage contents do not match')
+
+    def test_base_download_with_capitalise_small_caps(self):
+        download_path = './tmp/test_base_download_with_capitalise_small_caps/'
+        # WEB doesn't have any passages that apply "small caps" styling
+        bible = BaseDownloader(file_writing_function=yaml_file_interface.write, default_directory=download_path,
+                               translation='KJV', capitalise_small_caps=True)
+        bible.download_passage('Isaiah', 42, 5)
+        downloaded_file = yaml_file_interface.read(f'{download_path}/Isaiah')
+        static_file_path = f'{self.get_test_directory("KJV")}/test_base_download_with_capitalise_small_caps.yaml'
+        static_file = yaml_file_interface.read(static_file_path)
+        self.assertEqual(downloaded_file['Isaiah'], static_file['Isaiah'], 'Passage contents do not match')
+
+    def test_base_download_with_get_raw_output(self):
+        download_path = './tmp/test_base_download_with_get_raw_output/'
+        bible = BaseDownloader(file_writing_function=yaml_file_interface.write, default_directory=download_path)
+        bible.download_chapter('1 John', 1)
+        downloaded_file = yaml_file_interface.read(f'{download_path}/1 John')
+        # Raw output should be the same as the extractor's output after writing the equivalent file contents
+        bible.get_raw_output = True
+        dry_run_contents = bible.download_chapter('1 John', 1)
+        self.assertEqual(downloaded_file['1 John'], dry_run_contents['1 John'],
+                         'Incorrect number of passages downloaded')
 
 
 if __name__ == "__main__":